requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
cryptography==2.4.1
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
passlib==1.7.1
pycparser==2.19
pylint==2.1.1
python-dotenv==0.10.0
six==1.11.0
SQLAlchemy==1.2.14
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
